fix(openshift): convert Route weight and targetPort to float64 for unstructured object compatibility
- Cast weight value from int to float64 in Route spec
- Cast targetPort value from int to float64 in Route port configuration
- Fix whitespace indentation for route variable declaration

+++ internal/resources/route.go
@@ -38,7 +38,7 @@ func buildRoute(name, namespace, serviceName string, labels map[string]string, a
 		annotationsMap[k] = v
 	}
 
-	route := &unstructured.Unstructured{
+		route := &unstructured.Unstructured{
 		Object: map[string]interface{}{
 			"apiVersion": "route.openshift.io/v1",
 			"kind":       "Route",
@@ -52,10 +52,10 @@ func buildRoute(name, namespace, serviceName string, labels map[string]string, a
 				"to": map[string]interface{}{
 					"kind":   "Service",
 					"name":   serviceName,
-					"weight": 100,
+					"weight": float64(100),
 				},
 				"port": map[string]interface{}{
-					"targetPort": targetPort,
+					"targetPort": float64(targetPort),
 				},
 				"path": path,
 			},
